refactored auth tests
